Use Arc more and other changes.
